requirements.txt
google-generativeai>=0.5.0
supabase>=2.0.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
reportlab>=4.1.0
python-multipart>=0.0.9
PyPDF2>=3.0.0
